move scan transaction to sync_block
